selecting instantiate function is easier
